Cleared up location labelling
Locations are now labelled correctly. There is still a problem that if
there is a single point at the colony between trips, then that is
labelled a start, and not and end (i.e. it can't be both), so a small
proportion of trips (~10%) don't have a final end point at the colony,
i.e. their final point is that before they enter the 0.5 km readius
around their nest.

This commit also contains some debug code used to clear this up - this
will be removed for the next commit.